departmentheads apllications can not be signed by other department heads
department heads may now only sign applications of members of the department who are not department heads of the department. If a department with multiplw heads have some rank between them, so a new department have to be created "headsDepartment" in wich member are heads of original department and the head is the member of highest rank.